django/requirements/base.txt
Django>=4.2,<5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
whitenoise==6.6.0
psutil==5.9.6
Pillow==10.4.0
